chore(refresh): Tier 0 cleanup — analyze.sh + 7 warning fixes + Makefile floor
Cleanup pass surfaced by a fresh code review + new scripts/analyze.sh
static-analysis script (cppcheck + clang -Wextra -Wshadow). All
behaviour-preserving; verified fps-neutral on the two reachable bench
machines:

  mini-intel demo1 1024x768: 101.20 fps (unchanged from 78c26f2 row)
  yosemite   demo1 1024x768:  25.00 fps (unchanged from e8ae174 row)

mini-g4 not benched this round — the box's Radeon 9200 reports an
empty Displays: list (system_profiler SPDisplaysDataType), so SDL
can't create a GL pixel format. Hardware/wiring issue, unrelated to
this change.

Fixes:
  - header/local.h:453: R_DrawParticles2 colortable[768] → [256] to
    match the definition. C decays array params to pointers so the
    runtime was always fine, but clang -Warray-parameter flagged it.
  - r_surf.c:896: R_RenderLightmappedPoly stack temp[128*128] → [34*34]
    to match R_RenderBrushPoly's identical buffer above. Saves 63 KB
    of stack walk per dynamic surface per frame.
  - r_mesh.c R_LerpVerts: explicitly zero lerp[3] in both branches so
    the vec4_t's 4th lane isn't a read-of-uninit when downstream
    code (R_DrawAliasShadow memcpy) reads it as a vec3_t.
  - r_mesh.c R_DrawAliasFrameLerp: drop the inner `float l` redecl
    inside the gl_vertex_arrays block — it shadowed the outer `l`
    at line 92 with identical semantics.
  - r_misc.c SCR_ScreenShot_f: guard both malloc() calls against
    OOM-NULL before pointer arith / memset.
  - r_light.c R_BuildLightMap: cast `size` to size_t for the
    sizeof-shift comparison.

Tooling:
  - scripts/analyze.sh: cppcheck warning class + clang -Wextra
    -Wshadow -Wundef -Wpointer-arith. Filters stb_image.h and the
    -Wmissing-prototypes flood from file-local funcs that should be
    static (separate cleanup). Always exit 0 — developer tool, not
    a CI gate.
  - Makefile CFLAGS += -Wpointer-arith. Tight enough to catch real
    bugs without flooding build logs with the K&R-zlib and client/
    -Wshadow noise.

Also lands the rewritten NEXT_ROUND_PLAN.md derived from the fresh
code review (see commit body for rationale).

+++ yquake2/src/refresh/r_light.c
@@ -498,7 +498,7 @@ R_BuildLightMap(msurface_t *surf, byte *dest, int stride)
 	tmax = (surf->extents[1] >> 4) + 1;
 	size = smax * tmax;
 
-	if (size > (sizeof(s_blocklights) >> 4))
+	if ((size_t)size > (sizeof(s_blocklights) >> 4))
 	{
 		ri.Sys_Error(ERR_DROP, "Bad s_blocklights size");
 	}
